simulator/simulator.go
package simulator

import (
	"context"
	"fmt"
	"math/big"
	"strings"
	"time"

	"github.com/ethereum/go-ethereum"
	"github.com/ethereum/go-ethereum/accounts/abi"
	"github.com/ethereum/go-ethereum/common"
	"github.com/ethereum/go-ethereum/core/types"
	"github.com/ethereum/go-ethereum/ethclient"
	"github.com/michaelpento.lv/mevbot/types"
)

// SimulationResult represents the result of a transaction simulation
type SimulationResult struct {
	Success bool
	GasUsed uint64
	Error   error
}

// Simulator handles transaction simulation
type Simulator struct {
	client *ethclient.Client
}

// NewSimulator creates a new transaction simulator
func NewSimulator(client *ethclient.Client) *Simulator {
	return &Simulator{
		client: client,
	}
}

// SimulateTransaction simulates a transaction
func (s *Simulator) SimulateTransaction(ctx context.Context, tx *types.Transaction) (*SimulationResult, error) {
	// Get current gas price
	gasPrice, err := s.client.SuggestGasPrice(ctx)
	if err != nil {
		return nil, fmt.Errorf("failed to get gas price: %w", err)
	}

	// Simulate the transaction
	gasUsed, err := s.client.EstimateGas(ctx, ethereum.CallMsg{
		From:     tx.From(),
		To:       tx.To(),
		Gas:      tx.Gas(),
		GasPrice: gasPrice,
		Value:    tx.Value(),
		Data:     tx.Data(),
	})
	if err != nil {
		return &SimulationResult{
			Success: false,
			Error:   err,
			GasUsed: gasUsed,
		}, nil
	}

	// Try executing the call
	result, err := s.client.CallContract(ctx, ethereum.CallMsg{
		From:     tx.From(),
		To:       tx.To(),
		Gas:      tx.Gas(),
		GasPrice: gasPrice,
		Value:    tx.Value(),
		Data:     tx.Data(),
	}, nil)
	if err != nil {
		return &SimulationResult{
			Success: false,
			Error:   err,
			GasUsed: gasUsed,
		}, nil
	}

	return &SimulationResult{
		Success: true,
		GasUsed: gasUsed,
	}, nil
}

// SimulateFlashLoan simulates a flash loan
func (s *Simulator) SimulateFlashLoan(ctx context.Context, token common.Address, amount *big.Int) (*SimulationResult, error) {
	// Get current gas price
	gasPrice, err := s.client.SuggestGasPrice(ctx)
	if err != nil {
		return nil, fmt.Errorf("failed to get gas price: %w", err)
	}

	// Create flash loan call data
	callData, err := s.createFlashLoanCallData(token, amount)
	if err != nil {
		return nil, fmt.Errorf("failed to create flash loan call data: %w", err)
	}

	// Simulate the transaction
	gasUsed, err := s.client.EstimateGas(ctx, ethereum.CallMsg{
		From:     common.Address{}, // Will be set by node
		To:       &token,
		Gas:      2000000, // Higher gas limit for flash loans
		GasPrice: gasPrice,
		Value:    big.NewInt(0),
		Data:     callData,
	})
	if err != nil {
		return &SimulationResult{
			Success: false,
			Error:   err,
			GasUsed: gasUsed,
		}, nil
	}

	// Try executing the call
	result, err := s.client.CallContract(ctx, ethereum.CallMsg{
		From:     common.Address{}, // Will be set by node
		To:       &token,
		Gas:      2000000, // Higher gas limit for flash loans
		GasPrice: gasPrice,
		Value:    big.NewInt(0),
		Data:     callData,
	}, nil)
	if err != nil {
		return &SimulationResult{
			Success: false,
			Error:   err,
			GasUsed: gasUsed,
		}, nil
	}

	return &SimulationResult{
		Success: true,
		GasUsed: gasUsed,
	}, nil
}

// createFlashLoanCallData creates the call data for a flash loan
func (s *Simulator) createFlashLoanCallData(token common.Address, amount *big.Int) ([]byte, error) {
	// Create flash loan pool ABI
	poolABI, err := abi.JSON(strings.NewReader(FlashLoanPoolABI))
	if err != nil {
		return nil, fmt.Errorf("failed to parse pool ABI: %w", err)
	}

	// Pack the flash loan function call
	data, err := poolABI.Pack(
		"flashLoan",
		common.Address{}, // Will be set by node
		token,
		amount,
		[]byte{}, // Callback data will be set by node
	)
	if err != nil {
		return nil, fmt.Errorf("failed to pack flash loan function call: %w", err)
	}

	return data, nil
}

// SimulateArbitrage simulates an arbitrage trade
func (s *Simulator) SimulateArbitrage(ctx context.Context, opp *types.ArbitrageOpportunity) (*SimulationResult, error) {
	// Get current gas price
	gasPrice, err := s.client.SuggestGasPrice(ctx)
	if err != nil {
		return nil, fmt.Errorf("failed to get gas price: %w", err)
	}

	// Create simulation call
	callData, err := s.createArbitrageCallData(opp)
	if err != nil {
		return nil, fmt.Errorf("failed to create call data: %w", err)
	}

	// Simulate the transaction
	gasUsed, err := s.client.EstimateGas(ctx, ethereum.CallMsg{
		From:     common.Address{}, // Will be set by node
		To:       &opp.Exchanges[0].GetRouterAddress(),
		Gas:      1000000, // Conservative estimate
		GasPrice: gasPrice,
		Value:    big.NewInt(0),
		Data:     callData,
	})
	if err != nil {
		return &SimulationResult{
			Success: false,
			Error:   err,
			GasUsed: gasUsed,
		}, nil
	}

	// Try executing the call
	result, err := s.client.CallContract(ctx, ethereum.CallMsg{
		From:     common.Address{}, // Will be set by node
		To:       &opp.Exchanges[0].GetRouterAddress(),
		Gas:      1000000, // Conservative estimate
		GasPrice: gasPrice,
		Value:    big.NewInt(0),
		Data:     callData,
	}, nil)
	if err != nil {
		return &SimulationResult{
			Success: false,
			Error:   err,
			GasUsed: gasUsed,
		}, nil
	}

	return &SimulationResult{
		Success: true,
		GasUsed: gasUsed,
	}, nil
}

// createArbitrageCallData creates the call data for an arbitrage trade
func (s *Simulator) createArbitrageCallData(opp *types.ArbitrageOpportunity) ([]byte, error) {
	// Create router ABI
	routerABI, err := abi.JSON(strings.NewReader(UniswapV2Router02ABI))
	if err != nil {
		return nil, fmt.Errorf("failed to parse router ABI: %w", err)
	}

	// Pack the swap function call
	deadline := big.NewInt(time.Now().Add(15 * time.Minute).Unix())
	data, err := routerABI.Pack(
		"swapExactTokensForTokens",
		opp.InputAmount,
		opp.OutputAmount,
		opp.Path,
		common.Address{}, // Will be set by node
		deadline,
	)
	if err != nil {
		return nil, fmt.Errorf("failed to pack function call: %w", err)
	}

	return data, nil
}
